Add export barrier types to interactions dataset.
